자신의 정보만 보기&수정 가능

+++ backend/src/main/java/com/example/backend/controller/member/MemberController.java
@@ -6,6 +6,8 @@ import com.example.backend.service.member.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -75,34 +77,53 @@ public class MemberController {
     }
 
     @GetMapping("{id}")
-    public Member getMember(@PathVariable String id) {
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Member> getMember(@PathVariable String id, Authentication auth) {
 
-        return service.get(id);
+        if (service.hasAccess(id, auth)) {
+            return ResponseEntity.ok(service.get(id));
+        } else {
+
+            return ResponseEntity.status(403).build();
+        }
+
     }
 
     @DeleteMapping("remove")
-    public ResponseEntity<Map<String, Object>> remove(@RequestBody Member member) {
-        if (service.remove(member)) {
-            //잘됨
-            return ResponseEntity.ok(Map.of("message",
-                    Map.of("type", "success", "text", "회원정보를 삭제하였습니다")));
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Map<String, Object>> remove(@RequestBody Member member, Authentication authentication) {
+        if (service.hasAccess(member.getId(), authentication)) {
 
+            if (service.remove(member)) {
+                //잘됨
+                return ResponseEntity.ok(Map.of("message",
+                        Map.of("type", "success", "text", "회원정보를 삭제하였습니다")));
+
+            } else {
+                return ResponseEntity.badRequest().body((Map.of("message",
+                        Map.of("type", "warning", "text", "정확한 정보를 입력해 주세요"))));
+            }
         } else {
-            return ResponseEntity.badRequest().body((Map.of("message",
-                    Map.of("type", "warning", "text", "정확한 정보를 입력해 주세요"))));
+            return ResponseEntity.status(403).build();
         }
     }
 
     @PutMapping("update")
-    public ResponseEntity<Map<String, Object>> update(@RequestBody MemberEdit member) {
-        if (service.update(member)) {
-            //잘됨
-            return ResponseEntity.ok(Map.of("message",
-                    Map.of("type", "success", "text", "회원정보를 변경하였습니다")));
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Map<String, Object>> update(@RequestBody MemberEdit member, Authentication authentication) {
+        if (service.hasAccess(member.getId(), authentication)) {
 
+            if (service.update(member)) {
+                //잘됨
+                return ResponseEntity.ok(Map.of("message",
+                        Map.of("type", "success", "text", "회원정보를 변경하였습니다")));
+
+            } else {
+                return ResponseEntity.badRequest().body((Map.of("message",
+                        Map.of("type", "warning", "text", "정확한 정보를 입력해 주세요"))));
+            }
         } else {
-            return ResponseEntity.badRequest().body((Map.of("message",
-                    Map.of("type", "warning", "text", "정확한 정보를 입력해 주세요"))));
+            return ResponseEntity.status(403).build();
         }
     }
